Extract events to separate files
