+++ mirai-core/src/commonMain/kotlin/net.mamoe.mirai/event/events/EventCancelledException.kt
@@ -1,0 +1,9 @@
+package net.mamoe.mirai.event.events
+
+@Suppress("unused")
+class EventCancelledException : RuntimeException {
+    constructor() : super()
+    constructor(message: String?) : super(message)
+    constructor(message: String?, cause: Throwable?) : super(message, cause)
+    constructor(cause: Throwable?) : super(cause)
+}
